Moving functions from main.py. Moving functions createAtoms, createBravaislattice, MSD, MSD_plot and self_diffusion_coefficient from main to separate files.
